Prism Algorithm
Due to global climate change in recent years and the impact of EI Nino, the occurrence frequency of natural disasters in various places has continued to rise.
These disasters also exhibit ever greater variety, complexity and negatively impact human beings’ lives and property to a greater degree.
Among them, the threats to isolated areas are the most profound. Once an isolated area is struck by a disaster whereby key facilities are damaged, the community will bear the impact of the islanding effect, that is, materials and medicines cannot be delivered timely, and with a damaged healthcare system, the lives of people will be threatened.